User History 1: City Creation

+++ src/main/java/cujae/edmaps/core/Route.java
@@ -2,9 +2,9 @@ package cujae.edmaps.core;
 
 public class Route {
     private Bus bus;
-    private Double distance;
+    private Float distance;
 
-    public Route(Bus bus, Double distance) {
+    public Route(Bus bus, Float distance) {
         setBus(bus);
         setDistance(distance);
     }
@@ -17,11 +17,11 @@ public class Route {
         this.bus = bus;
     }
 
-    public Double getDistance() {
+    public Float getDistance() {
         return distance;
     }
 
-    public boolean setDistance(Double distance) {
+    public boolean setDistance(Float distance) {
         if (distance < 0) return false;
         else this.distance = distance;
         return true;
